pyproject.toml
[tool.poetry]
name = "unifiapipy"
version = "0.1.0"
description = ""
authors = ["Quinn Batten <qbatten@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "unifiapipy", from = "src"}]

[tool.poetry.scripts]
scrape_all = "unifiapipy.main:do_full_scrape"
get_aps = "unifiapipy.main:get_aps"

[tool.poetry.dependencies]
python = "^3.10"
requests = "~2"
plotly = "~5"
pandas = "^1.5.1"
boto3 = "^1.26.26"
click = "^8.1.3"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
black = "~22"
pytest = "~7"
pre-commit = "^2.20.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
